3e63388 fix the size_t vs int plus a few I missed

+++ smtk/Qt/qtUIManager.cxx
@@ -661,7 +661,7 @@ QWidget* qtUIManager::createComboBox(
     {
     std::string enumText = itemDef->discreteEnum(i);
     if(itemDef->hasDefault() &&
-      itemDef->defaultDiscreteIndex() == i)
+      static_cast<size_t>(itemDef->defaultDiscreteIndex()) == i)
       {
       tooltip = "Default: " + QString(enumText.c_str());
       }
